Refactor lib/utils module and update Installer class
The utility functions in lib/utils have been refactored and relocated into different files. The installer class dependency paths in lib/Installer.js and its constructor have also been updated. Consequently, the bootstrap fields in Installer now depend on an externally provided 'paths' object.
