+++ lib/utils/data.js
@@ -1,7 +1,8 @@
-const path = require('path');
+
 const fs = require('fs');
-const { exec, child_process} = require('node:child_process');
 const {get} = require("axios");
+const {pkgFileName} = require("../constants");
+const Initialize = require("../Initialize");
 
 /**
  * Reads a PHP file, extracts, and returns the WordPress version number.
@@ -22,53 +23,42 @@ function geVarFromPHPFile (fileContent, variableName = '$wp_version') {
 }
 
 /**
- * Returns the locale of the user.
- *
- * @returns {string} The locale of the user.
- */
-function getUserLocale() {
-  return Intl.DateTimeFormat().resolvedOptions().locale;
-}
-
-/**
  * Reads wp-package.json from the root folder and extracts the value of the --template option or uses the default.
  * The default config is used if no template is provided. Checks if the template file exists and reads it as JSON if it does.
  *
  * @param {any} args - The arguments object.
- * @param {string} args.template - the path to the template file
+ * @param {string} args.template - The path to the template file.
  * @return {WPMMconfig} The configuration object.
  */
 function getConfig (args) {
   /**
    * The default config from the root plugin folder. This is used if no template is provided
-   * @type {WPMMconfig} defaultConfig - The default configuration object
+   *
+   * @type {WPMMconfig} config - The configuration object
    */
-  const defaultConfig = JSON.parse(fs.readFileSync(path.join(__dirname, '..','..', 'wp-package.json'), 'utf8'));
-
   let config = {};
 
   // Extract the value of the --template option or use the default
   if (! args?.template) {
-    const templatePath = 'wp-package.json';
+    const templatePath = pkgFileName;
 
     // Check if the template file exists and read it as JSON if it does
     if (!fs.existsSync(templatePath)) {
-      console.error(`🔴 The template file ${templatePath} does not exist.`);
+      console.log(`🔴 The template file ${templatePath} does not exist.`);
+      const init = new Initialize();
+      config = init.generateConfig();
     } else {
-      /**
-       * Read the user-defined configuration object from the template file and add it to the config
-       * @type {WPMMconfig} config - The configuration object
-       */
-      config = {...defaultConfig, ...JSON.parse(fs.readFileSync(templatePath, 'utf8'))};
+      config = JSON.parse(fs.readFileSync(templatePath, 'utf8'));
+      console.log(`🔴 The template file ${templatePath} exists in the root folder. Using it.`, config);
     }
+
   } else {
     /**
      * the user has provided a template file via the --template option. read the config from the remote source
      */
     get(args.template)
       .then(response => {
-        const gitConfig = response.data;
-        config = {...defaultConfig, ...gitConfig};
+        config = response.data;
       })
       .catch(error => {
         console.log("🔴 Error: " + error.message);
@@ -137,94 +127,6 @@ function getDownloadUrl (packageName, packageVersion, type) {
 }
 
 /**
- * Installs npm packages in the specified package directory.
- *
- * @param {string} packageDirectory - The directory path where the package is located.
- * @return {Promise<void>} - A promise that resolves when the packages are installed and built.
- */
-async function installNpmPackages (packageDirectory) {
-  const packageJsonPath = path.join(packageDirectory, 'package.json');
-  if (!fs.existsSync(packageJsonPath)) {
-    // console.warn(`Local directory (${packageDirectory}) does not contain a package.json file.`)
-    return;
-  }
-  const packageData = JSON.parse(fs.readFileSync(packageJsonPath, 'utf8'));
-  const packageName = packageData.name;
-  console.log(`🚀 Installing and building ${packageName} from local directory...`);
-
-  const packageLockPath = path.join(packageDirectory, 'package-lock.json');
-  let command = 'npm i && npm run build';
-  if (fs.existsSync(packageLockPath)) {
-    command = 'npm ci && npm run build';
-  }
-
-  await new Promise((resolve, reject) => {
-    exec(command, { cwd: packageDirectory }, (error) => {
-      if (error) {
-        reject(error);
-      } else {
-        console.log(`📦 ${packageName} dependencies installed and built.`);
-        resolve();
-      }
-    });
-  });
-}
-
-/**
- * Installs composer dependencies and generates autoloader based on composer.json file.
- *
- * @param {string} repoPath - The path to the repository where composer.json is located.
- *
- * @returns {Promise} - A promise resolving when the installation process is completed.
- */
-async function installComposer (repoPath) {
-  console.log('🎻 Found composer.json');
-  await exec('composer install --no-dev', { cwd: repoPath });
-  await exec('composer dumpautoload -o', { cwd: repoPath });
-}
-
-/**
- * Checks if WP-CLI is available.
- *
- * @return {boolean} - A promise that resolves to true if WP-CLI is available, false otherwise.
- */
-async function isWPCLIAvailable () {
-  try {
-    // Attempt to execute a simple WP-CLI command
-    await exec('wp --version');
-    return true; // If successful, WP-CLI is available
-  } catch (error) {
-    console.log('🔴 WP-CLI is not available on this system. Please install WP-CLI and try again.');
-    console.log('Read more about at https://make.wordpress.org/cli/handbook/guides/installing/');
-    return false; // If an error occurs, WP-CLI is not available
-  }
-}
-
-/**
- * Runs post-install commands asynchronously.
- *
- * @param {Array} commands - An array of WP-CLI commands to execute.
- * @return {Promise<void>} - A promise that resolves when the post-install commands complete.
- */
-async function runPostInstallCommands (commands) {
-  // Execute each post-install command
-  for (const command of commands) {
-    try {
-      console.log(`Executing: ${command}`);
-      const { stdout, stderr } = await child_process.exec(command);
-      if (stdout) {
-        console.log(`Command output:\n${stdout}`);
-      }
-      if (stderr) {
-        console.error(`Command error:\n${stderr}`);
-      }
-    } catch (error) {
-      console.error(`Error executing command: ${command}`, error);
-    }
-  }
-}
-
-/**
  * Retrieve information about the WPMM and system environment.
  *
  * @return {void}
@@ -260,15 +162,10 @@ function printTimePassed (startTime) {
 
 module.exports = {
   geVarFromPHPFile,
-  getUserLocale,
   getConfig,
+  getInfo,
   getConnectionSettings,
   getWordPressDownloadUrl,
   getDownloadUrl,
-  getInfo,
-  installNpmPackages,
-  installComposer,
-  isWPCLIAvailable,
-  runPostInstallCommands,
   printTimePassed
 };
