+++ lib/utils/wpConfig.js
@@ -1,8 +1,9 @@
 const path = require("path");
 const fs = require("fs");
-const {geVarFromPHPFile} = require("./index");
+const {geVarFromPHPFile} = require("./data.js");
 const axios = require("axios");
-const {defaultWpInstallFolder} = require("../constants");
+const {defaultWpInstallFolder} = require("../constants.js");
+const {defaultWpInstallLanguage} = require("../constants");
 
 /**
  * Gets the default paths for the WordPress installation.
@@ -11,16 +12,15 @@ const {defaultWpInstallFolder} = require("../constants");
  * @return {WPMMpaths} - An object containing the default paths for the WordPress installation.
  */
 function getWordPressPaths (config, rootFolder = process.cwd()) {
+  const baseFolder = path.join(rootFolder, config.name ?? defaultWpInstallFolder);
   return {
     tempDir: path.join(rootFolder, 'temp'),
     rootFolder: rootFolder,
-    baseFolder: path.join(rootFolder, config.name ?? defaultWpInstallFolder),
-    pluginsFolder: path.join(this.baseFolder, 'wp-content', 'plugins'),
-    themeFolder: path.join(this.baseFolder, 'wp-content', 'themes')
+    baseFolder: baseFolder,
+    pluginsFolder: path.join(baseFolder, 'wp-content', 'plugins'),
+    themeFolder: path.join(baseFolder, 'wp-content', 'themes')
   };
 }
-
-
 
 /**
  * Fetch the WordPress version from WordPress.org
@@ -28,15 +28,23 @@ function getWordPressPaths (config, rootFolder = process.cwd()) {
  * @async
  * @return {string} the version of WordPress
  */
-async function getLastWpVersion () {
-  const response = await axios.get('https://api.wordpress.org/core/version-check/1.7/')
-    .then(() => {
-      if (response.data?.offers && response.data.offers.length > 0) {
-        return response.data.offers[0];
-      } else {
-        throw new Error('❗Cannot get the last version available');
-      }
-    });
+async function getLastWp () {
+  const response = await axios.get('https://api.wordpress.org/core/version-check/1.7/');
+
+  if (response.data?.offers && response.data.offers.length > 0) {
+    return response.data.offers[0];
+  } else {
+    throw new Error('❗Cannot get the last version available');
+  }
+}
+
+/**
+ * Returns the locale of the user.
+ *
+ * @returns {string} The locale of the user.
+ */
+function getUserLocale() {
+  return Intl.DateTimeFormat().resolvedOptions().locale || defaultWpInstallLanguage;
 }
 
 /**
@@ -116,7 +124,8 @@ function replaceEmptySalts (configContent) {
 }
 
 module.exports = {
-  getLastWpVersion,
+  getLastWp,
+  getUserLocale,
   getWordPressPaths,
   getCurrentWpInfo,
   replaceDbConstant,
